Add rentalsmodules.rb>> update the app.rb
